requirements.txt
fastapi
uvicorn
jinja2
pymongo
python-dotenv #store your mongo uri 